requirements.txt
pydriller
convex
python-dotenv
langchain
langchain-openai
pydantic
tiktoken